requirements.txt
numpy==1.9.3
Pillow==2.9.0
wheel==0.24.0
